+++ CHANGELOG.md
@@ -5,10 +5,12 @@
 ### Features
 
 * Don't display webchat when viewing a playlist.
+* Better default values for settings.
 
 ### Fixes
 
 * Fix: starting with Peertube 3.2.0, there is a header 'X-Frame-Options'. Removing it on the iframe route.
+* Remove old Peertube 3.0.1 compatibility (default value for checkbox settings should be false)
 
 ## v2.1.3